revert public API break, keep rewriter in dynamoCore but in new file
